performance improvements

+++ tasks/support/package_worker.rb
@@ -11,13 +11,14 @@ module Tasks
 
       def process_package(package_mapping, path)
         LOGGER.info "Processing ... #{package_mapping}"
-        Helpers::HeavyFileUtils.unpack_tar_gz(Helpers::HeavyFileUtils.download(package_mapping[url]))
-        desc_path = "#{path}/#{DESC_FILE_NAME}"
-        desc = DebControl::ControlFileBase.read(desc_path).force_encoding('UTF-8')
+        Helpers::HeavyFileUtils.unpack_tar_gz(Helpers::HeavyFileDownloader.download(package_mapping[url]))
+        desc = DebControl::ControlFileBase.read("#{path}/#{DESC_FILE_NAME}")
         upsert_package_data(desc, package_mapping[url])
         LOGGER.info "Finished processing #{package_mapping}"
+      rescue StandardError => e
+        raise StandardError, "Mapping: #{package_mapping}\nMessage: #{e}"
       ensure
-        FileUtils.remove_dir(path.to_s)
+        FileUtils.remove_dir(path.to_s) if File.exist?(path.to_s)
       end
 
       def upsert_package_data(desc, url)
